Implement routing with Home and Leaderboard components; refactor App component
